Entities connected to the users

+++ MicromouseSimulatorBackend/MicromouseSimulatorBackend.BLL/Services/SimulationService.cs
@@ -33,18 +33,18 @@ namespace MicromouseSimulatorBackend.BLL.Services
             _simulationDockerService = simulationDockerService;
         }
 
-        public IEnumerable<SimulationExpanded> FindAll()
+        public IEnumerable<SimulationExpanded> FindAll(string userId)
         {
-            var result = _simulationRepository.FindAllAndPopulate().ToList();
+            var result = _simulationRepository.FindAllAndPopulate(userId).ToList();
             return result;
         }
 
-        public SimulationExpanded FindById(string id)
+        public SimulationExpanded FindById(string id, string userId)
         {
-            return _simulationRepository.FindByIdAndPopulate(id);
+            return _simulationRepository.FindByIdAndPopulate(id, userId);
         }
 
-        public Simulation Create(Simulation simulation)
+        public Simulation Create(Simulation simulation, string userId)
         {
             if (simulation.AlgorithmId != null && _algorithmRepository.FindById(simulation.AlgorithmId) == null)
                 throw new DocumentDoesntExistsException("The given AlgorithmId does not exist!");
@@ -53,12 +53,14 @@ namespace MicromouseSimulatorBackend.BLL.Services
             if (simulation.MouseId != null && _mouseRepository.FindById(simulation.MouseId) == null)
                 throw new DocumentDoesntExistsException("The given MouseId does not exist!");
 
+            simulation.UserId = userId;
             _simulationRepository.InsertOne(simulation);
             return simulation;
         }
-        public void Update(string id, Simulation simulation)
+
+        public void Update(string id, Simulation simulation, string userId)
         {
-            if (_simulationRepository.FindById(id) == null)
+            if (!simulationExists(id, userId))
                 throw new DocumentDoesntExistsException("No Simulation exists with the given ID!");
             if (simulation.AlgorithmId != null && _algorithmRepository.FindById(simulation.AlgorithmId) == null)
                 throw new DocumentDoesntExistsException("The given AlgorithmId does not exist!");
@@ -68,19 +70,22 @@ namespace MicromouseSimulatorBackend.BLL.Services
                 throw new DocumentDoesntExistsException("The given MouseId does not exist!");
 
             simulation.Id = id;
+            simulation.UserId = userId;
             _simulationRepository.ReplaceOne(id, simulation);
         }
 
-        public void Delete(string id)
+        public void Delete(string id, string userId)
         {
-            _simulationRepository.DeleteById(id);
-
-            _simulationFileService.DeleteById(id);
+            if (simulationExists(id, userId))
+            {
+                _simulationRepository.DeleteById(id);
+                _simulationFileService.DeleteById(id);
+            }
         }
 
-        public async Task<SimulationResult> RunSimulationAsync(string id)
+        public async Task<SimulationResult> RunSimulationAsync(string id, string userId)
         {
-            var simulation = _simulationRepository.FindByIdAndPopulate(id);
+            var simulation = _simulationRepository.FindByIdAndPopulate(id, userId);
             if (simulation == null)
                 throw new DocumentDoesntExistsException("No Simulation exists with the given ID!");
             if (simulation.Algorithm == null)
@@ -91,6 +96,16 @@ namespace MicromouseSimulatorBackend.BLL.Services
             var folderPath = _simulationFileService.Save(simulation);
             await _simulationDockerService.RunContainerAsync(simulation, folderPath);
             return _simulationFileService.ReadResult(simulation);
+        }
+
+        private bool simulationExists(string id, string userId)
+        {
+            if (!_simulationRepository.IsValidId(id))
+                return false;
+            var result = _simulationRepository.FilterBy(doc => doc.UserId == userId && doc.Id == id).FirstOrDefault();
+            if (result == null)
+                return false;
+            return true;
         }
     }
 }
